Exc 8: Fix RNG, turnaround time
This fixes a bug in the scheduler_rand() function by rounding the result
before casting (thus truncating) to int. Additionally, the reported
turnaround time should now be in line with the definition presented in
the lecture. Previously it was computed based on a process' start time,
instead of its arrival time.

The example output is updated to reflect the changes made.